+++ lib/base.rb
@@ -32,11 +32,12 @@ module Input
       @x = params[:x] || 0
       @y = params[:y] || 0
 
-      @font = params[:font].to_s
-      @size_enum = SIZE_ENUM.fetch(params[:size_enum] || :normal, :size_enum)
+      font = params[:font].to_s
+      size_enum = SIZE_ENUM.fetch(params[:size_enum] || :normal, params[:size_enum])
 
       word_chars = (params[:word_chars] || ('a'..'z').to_a + ('A'..'Z').to_a + ('0'..'9').to_a + ['_', '-'])
-      _, @font_height = $gtk.calcstringbox(word_chars.join(''), @size_enum, @font)
+      @font_style = FontStyle.new(font: font, size_enum: size_enum, word_chars: word_chars)
+      @font_height = @font_style.font_height
       @word_chars = Hash[word_chars.map { [_1, true] }]
       @punctuation_chars = Hash[(params[:punctuation_chars] || %w[! % , . ; : ' " ` ) \] } * &]).map { [_1, true] }]
       @crlf_chars = { "\r" => true, "\n" => true }