+++ lib/input.rb
@@ -1,6 +1,7 @@
 # Initially based loosely on code from Zif (https://github.com/danhealy/dragonruby-zif)
 require_relative 'value.rb'
 require_relative 'line_collection.rb'
+require_relative 'font_style.rb'
 require_relative 'base.rb'
 require_relative 'text.rb'
 require_relative 'multiline.rb'